add PoolPct to ContestMetadata. Rename OneAuditContest to OneAuditContestInfo. Move CardPoolB, addOAClcaAssortersXX() to core.CardPool
