+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/core/Cvr.kt
@@ -1,6 +1,7 @@
 package org.cryptobiotic.rlauxe.core
 
 import org.cryptobiotic.rlauxe.audit.AuditableCard
+import org.cryptobiotic.rlauxe.oneaudit.unpooled
 
 // core abstraction for both CVR and MVR
 // assumes that a vote is 0 or 1.
@@ -112,10 +113,6 @@ data class CvrExport(val id: String, val group: Int, val votes: Map<Int, IntArra
         result = 31 * result + id.hashCode()
         votes.forEach { (contestId, candidates) -> result = 31 * result + contestId.hashCode() + candidates.contentHashCode() }
         return result
-    }
-
-    companion object {
-        const val unpooled = "unpooled"
     }
 
 }
